# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "BatchResponse": {
        "properties": {
          "accepted": {
            "description": "Events stored by this request.",
            "title": "Accepted",
            "type": "integer"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "conflicts": {
            "description": "Events whose id exists with different content (kept first).",
            "title": "Conflicts",
            "type": "integer"
          },
          "duplicates": {
            "description": "Events already stored with identical content.",
            "title": "Duplicates",
            "type": "integer"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/EventErrorOut"
            },
            "title": "Errors",
            "type": "array"
          },
          "rejected": {
            "description": "Events that were not stored; see `errors`.",
            "title": "Rejected",
            "type": "integer"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "accepted",
          "duplicates",
          "conflicts",
          "rejected",
          "errors",
          "server_time",
          "request_id"
        ],
        "title": "BatchResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ErrorCategory"
          },
          "code": {
            "description": "Stable machine-readable code, e.g. RUN_NOT_FOUND.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Also in the X-Request-ID response header.",
            "title": "Request Id"
          },
          "retryable": {
            "description": "Whether retrying the same request may succeed.",
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "category",
          "retryable",
          "request_id",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCategory": {
        "enum": [
          "AUTHENTICATION",
          "AUTHORIZATION",
          "VALIDATION",
          "RATE_LIMIT",
          "DEPENDENCY",
          "TIMEOUT",
          "CONFLICT",
          "NOT_FOUND",
          "POLICY",
          "INTERNAL"
        ],
        "title": "ErrorCategory",
        "type": "string"
      },
      "ErrorEnvelope": {
        "description": "The one error shape of every non-2xx response (spec §101.2).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "EventErrorOut": {
        "description": "Why one event of a batch was not stored.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The event's id, when it was readable.",
            "title": "Event Id"
          },
          "index": {
            "description": "Position of the event in the submitted `events` array.",
            "title": "Index",
            "type": "integer"
          },
          "issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IssueOut"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "index",
          "event_id",
          "code"
        ],
        "title": "EventErrorOut",
        "type": "object"
      },
...
      "EventResponse": {
        "properties": {
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          },
          "status": {
            "enum": [
              "accepted",
              "duplicate",
              "conflict"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "status",
          "server_time",
          "request_id"
        ],
        "title": "EventResponse",
        "type": "object"
      },
...
      "IssueOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "code",
          "message"
        ],
        "title": "IssueOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Project API key. Create one with `python -m abb_api.cli create-key`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Agent Black Box ingestion and query API. Authenticate with a project API key: `Authorization: Bearer abb_live_<key_id>.<secret>`.",
    "title": "Agent Black Box API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/events": {
      "post": {
        "operationId": "ingest_event_v1_events_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed batch or unsupported schema version."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed, unknown, revoked or expired API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The key lacks `events:write` or is not bound to a project."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Body (compressed or decompressed) or event too large."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not application/json, or an unsupported Content-Encoding."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The event is invalid (single-event endpoint)."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded; honour `Retry-After`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency is unavailable; retry with backoff."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Ingest a single event",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/v1/events/batch": {
      "post": {
        "description": "Accepts `{batch_id?, sent_at?, events: [...]}` (at most 1000 events, 5 MiB, optionally gzip). `202` means the valid events are committed. Invalid events are reported per event in `errors` and do not affect the others. Retrying a batch is safe.",
        "operationId": "ingest_batch_v1_events_batch_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed batch or unsupported schema version."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed, unknown, revoked or expired API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The key lacks `events:write` or is not bound to a project."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Body (compressed or decompressed) or event too large."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not application/json, or an unsupported Content-Encoding."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The event is invalid (single-event endpoint)."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded; honour `Retry-After`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency is unavailable; retry with backoff."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Ingest a batch of events",
        "tags": [
          "ingestion"
        ]
      }
    },
...
  }
}